fix(types): resolve type annotations in engine.py and remove redundant float calls in detectors

+++ threatvision/detectors/crowd.py
@@ -39,7 +39,7 @@ class CrowdDetector(BaseDetector):
             xmax = max(p.box[2] for p in persons)
             ymax = max(p.box[3] for p in persons)
 
-            crowd_confidence = min(0.98, float(len(persons) / (self.density_threshold * 2)))
+            crowd_confidence = min(0.98, len(persons) / (self.density_threshold * 2))
 
             return self.filter_by_confidence(
                 [
